+++ streamlit_canary/components/progress.py
@@ -4,12 +4,14 @@ example:
     book = openpyxl.load_workbook('test.xlsx')
     with progress('Processing sheets...', len(book.sheets)) as prog:
         for sheet in book.sheets:
+            # you can put this before or after item processing.
             prog.update(sheet.title)
             ...
 """
 import streamlit as st
 import typing as tp
 from contextlib import contextmanager
+from time import sleep
 
 
 @contextmanager
@@ -21,13 +23,15 @@ def progress(
     prog = _Progress(label, total)
     yield prog
     if auto_close:
+        sleep(0.5)
         prog.close()
 
 
 class _Progress:
     def __init__(self, label: str, total: int = 0):
-        self.index = 0
         self._prog = st.progress(0, label)
+        self.total = total
+        self.index = 0
     
     def update(self, text: str = ''):
         self.index += 1