+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Add Streamlit-based tree select.
 - Use `__revision__` instead of `__version__` to mark state soured.
+- Refactor `sc.progress`
 - Return streamlit process and window process for `sc.run(..., subthread=True)`.
 
 ### 0.2.1 (2026-04-07)